Add files via upload
Decided which attribute was going to be played. Pushed the cards of the round into a new deck ready to be given to the winner. Got the attribute stats based on which attribute the round was based on. Next need to compare those attributes, find a highest value, check for duplicates and award cards to the winner.