Add files via upload
Fixed the issue where game was ending after the round was a draw. Game is now in a state where it will continue to run until there is a winner. Need to put a statement at the bottom to output the name of the winner and change bool to escape while loop and end the program.