Add files via upload
User is now directly involved in the game by entering their name, how many players they want to play against and which attribute of their card they want to put forward. Currently running into an error at the end of the game when the user wins. Program ends as planned when AI wins.