Add files via upload
Deck would only divide into a whole number of players. Example - with a 30 card deck, if players did not equal a number that was divisible by 30, cards would be missed out. Now if players is not a number divisible by 30, some players will have an extra card than others. However, all the cards will now be in use.